requirements.txt
langchain
python-dotenv
langchain-groq
pandas
langchain-experimental
tabulate
flask